experiment: add Lit version of date-picker month calendar
