+++ packages/date-picker/src/vaadin-date-picker-overlay-content-mixin.js
@@ -324,15 +324,13 @@ export const DatePickerOverlayContentMixin = (superClass) =>
     ) {
       if (calendars && calendars.length) {
         calendars.forEach((calendar) => {
-          calendar.setProperties({
-            i18n,
-            minDate,
-            maxDate,
-            focusedDate,
-            selectedDate,
-            showWeekNumbers,
-            ignoreTaps,
-          });
+          calendar.i18n = i18n;
+          calendar.minDate = minDate;
+          calendar.maxDate = maxDate;
+          calendar.focusedDate = focusedDate;
+          calendar.selectedDate = selectedDate;
+          calendar.showWeekNumbers = showWeekNumbers;
+          calendar.ignoreTaps = ignoreTaps;
 
           if (theme) {
             calendar.setAttribute('theme', theme);
