+++ packages/date-picker/test/month-calendar-lit.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-lit-month-calendar.js';
+import './month-calendar.common.js';
