+++ packages/date-picker/test/month-calendar-polymer.test.js
@@ -1,0 +1,2 @@
+import '../src/vaadin-month-calendar.js';
+import './month-calendar.common.js';
